Fixed bug with gallery layout on non-chromium browser
Previously, when the gallery was viewed on a non-chromium browser such as Firefox, the images would be shown in only one column. This does not follow the intended design of the gallery. The bug was fixed by explicitly specifying the maximum number of columns (4) that can be shown on the webpage.